bugninja_v2/requirements.txt
playwright==1.40.0
openai==1.12.0
python-dotenv==1.0.0 